requirements.txt
Django==2.1
django-postgres-copy==2.3.7
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
